+++ CHANGELOG.md
@@ -10,6 +10,8 @@ v0.28.1 (master)
 * Replace mt_rand with random_int in generateKey()
 * Directly redirect to the login page if the user is not authenticated
 * Rewrite the resolveChats function resolve and order properly the missing chats in the open_chats table
+* Add a boot() method to the Base widget class to allow tasks registration
+* Clear Chats widget cache during the night
 
 v0.28
 ---------------------------